pyproject.toml
[project]
name = "simikit"
version = "0.0.3"
description = "Image Similarity Toolkit in Python"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "loguru>=0.7.3",
    "pillow>=11.1.0",
    "pywavelets>=1.8.0",
    "scikit-learn>=1.6.1",
    "tomli>=2.2.1",
    "torch>=2.6.0",
    "transformers>=4.50.3",
]

[tool.uv]
build-constraint-dependencies = ["setuptools==60.0.0"]

[tool.ruff.lint]
preview = true
select = ["E", "F", "W", "I", "UP"]
exclude = ["tests", "tests/*", "demos/*"]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.ruff]
line-length = 120
target-version = "py310"


[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[dependency-groups]
ruff = [
    "pytest-cov>=6.0.0",
    "ruff>=0.11.2",
]
test = [
    "pytest>=8.3.5",
]
